Create order and blog routes. Overhaul role management & user model.
Had to scrap auth0 for role management as it was  discovered that it
 was a paid feature that is only temporarily available to free users.
Resorted to splitting the retailer and producer models and taking
the common properties of each and consolidating them into a
"user" model which specified the role. A middleware has been
implemented to attach the correct role to each request on receipt.

Initial post/get/patch routes in place for orders. Blog has post/get,
still need to implement patch/delete.