Notification handling adjusted to dismiss all notifications simultaneously.
Probably worth using the individual route for when a user manually
navigates to a page with a relevant notification.

Attached business name and image to the notification object
in the array on the user object in the main GET route.
Unfortunately the image format isn't particularly easy to render
on the front end. It might need another rework to place images
in directories instead.